New spotbugs fixes

+++ observation-manager/observation-manager-app/src/main/java/de/lehmannet/om/ui/navigation/tableModel/TableSorter.java
@@ -629,7 +629,7 @@ public class TableSorter extends AbstractSchemaTableModel {
             Color color = c == null ? Color.GRAY : c.getBackground();
             // In a compound sort, make each succesive triangle 20%
             // smaller than the previous one.
-            int dx = (int) (size / 2 * Math.pow(0.8, priority));
+            int dx = (int) ((double) size / 2 * Math.pow(0.8, priority));
             int dy = descending ? dx : -dx;
             // Align icon (roughly) with font baseline.
             y = y + 5 * size / 6 + (descending ? -dy : 0);
